functions/reliefweb/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.16.4
google-cloud-pubsub==2.18.4
functions-framework==3.5.0
Flask==3.0.0
requests==2.31.0
python-dateutil==2.8.2
